Crate wallet_service_mirror_public

The public side of wallet-service-mirror. This program opens two listening ports:

A GRPC server for receiving incoming poll requests from the private side of the mirror 2) An http(s) server for receiving client requests which will then be forwarded to the wallet service instance sitting behind the private part of the mirror.

Modules

mirror_service ğŸ”’

query ğŸ”’

Utility entity for managing queries submitted over our rocket endpoint and resolved by the GRPC polling mechanism.

utils ğŸ”’

Structs

BadRequest

Sets the status of the response to 400 (Bad Request).

ClientUriScheme

Wallet Service Mirror Uri Scheme

Config

Command line config

State ğŸ”’

State that is accessible by all rocket requests

Statics

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

Functions

build_grpc_server ğŸ”’

Starts the GRPC server in its own thread. This is necessary because the GRPC server does not implement Sync, which is a requirement to be managed by Rocket OR pass over an async await block.

build_rocket ğŸ”’

Builds the rocket instance given the configuration.

encrypted_request ğŸ”’

main ğŸ”’

monomorphized_function ğŸ”’

monomorphized_function ğŸ”’

unencrypted_request ğŸ”’

Type Definitions

ClientUri

Opaque Types

ğŸ”’

ğŸ”’